requirements.txt
autopep8==2.0.1
build==1.2.1
CacheControl==0.14.0
cachetools==5.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
crashtest==0.4.1
distlib==0.3.8
dulwich==0.21.7
fastjsonschema==2.19.1
filelock==3.14.0
google-api-core==2.19.1
google-auth==2.33.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
grpcio==1.65.4
grpcio-status==1.65.4
idna==3.7
importlib_metadata==7.1.0
installer==0.7.0
jaraco.classes==3.4.0
keyring==24.3.1
more-itertools==10.2.0
msgpack==1.0.8
packaging==24.0
pexpect==4.9.0
pkginfo==1.11.0
platformdirs==4.2.2
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
proto-plus==1.24.0
protobuf==5.27.3
ptyprocess==0.7.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.10.0
pycparser==2.22
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
rapidfuzz==3.9.3
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
shellingham==1.5.4
six==1.16.0
tomli==2.0.1
tomlkit==0.12.5
trove-classifiers==2024.5.22
urllib3==2.2.1
virtualenv==20.26.2
xattr==1.1.0
zipp==3.19.2
memory-profiler==0.61.0
